--- src/main/java/com/develhope/spring/AdminRepository.java
@@ -1,8 +1,0 @@
-package com.develhope.spring;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AdminRepository extends JpaRepository<Admin,Long> {
-}
