+++ src/main/java/com/develhope/spring/Order/OrderController.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Order;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class OrderController {
+}
