+++ src/main/java/com/develhope/spring/Order/OrderService.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Order;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderService {
+}
