+++ src/main/java/com/develhope/spring/Purchase/PurchaseController.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Purchase;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class PurchaseController {
+}
