+++ src/main/java/com/develhope/spring/Purchase/PurchaseRepository.java
@@ -1,5 +1,6 @@
-package com.develhope.spring;
+package com.develhope.spring.Purchase;
 
+import com.develhope.spring.Purchase.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
