+++ src/main/java/com/develhope/spring/Purchase/PurchaseService.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Purchase;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class PurchaseService {
+}
