+++ src/main/java/com/develhope/spring/Rent/RentController.java
@@ -1,0 +1,8 @@
+package com.develhope.spring.Rent;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+
+public class RentController {
+}
