+++ src/main/java/com/develhope/spring/Rent/RentRepository.java
@@ -1,8 +1,8 @@
-package com.develhope.spring;
+package com.develhope.spring.Rent;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SellerRepository extends JpaRepository<Seller, Long> {
+public interface RentRepository extends JpaRepository {
 }
