+++ src/main/java/com/develhope/spring/Rent/RentService.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Rent;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class RentService {
+}
