+++ src/main/java/com/develhope/spring/User/AdminService.java
@@ -1,5 +1,14 @@
-package com.develhope.spring;
+package com.develhope.spring.User;
 
+import com.develhope.spring.Order.Order;
+import com.develhope.spring.Order.OrderRepository;
+import com.develhope.spring.Purchase.Purchase;
+import com.develhope.spring.Purchase.PurchaseRepository;
+import com.develhope.spring.Rent.Rent;
+import com.develhope.spring.Rent.RentRepository;
+import com.develhope.spring.Vehicles.EntityofVehicles.Status;
+import com.develhope.spring.Vehicles.EntityofVehicles.Vehicles;
+import com.develhope.spring.Vehicles.VehiclesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
