+++ src/main/java/com/develhope/spring/User/SellerRepository.java
@@ -1,0 +1,9 @@
+package com.develhope.spring.User;
+
+import com.develhope.spring.User.Seller;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SellerRepository extends JpaRepository<Seller, Long> {
+}
