+++ src/main/java/com/develhope/spring/User/SellerService.java
@@ -1,4 +1,4 @@
-package com.develhope.spring;
+package com.develhope.spring.User;
 
 import org.springframework.stereotype.Service;
