+++ src/main/java/com/develhope/spring/User/corrette/UserController.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.User.corrette;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class UserController {
+}
