+++ src/main/java/com/develhope/spring/User/corrette/UserEnum.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.User.corrette;
+
+public enum UserEnum {
+    ADMIN,
+    SELLER,
+    CUSTOMER
+}
