+++ src/main/java/com/develhope/spring/User/corrette/UserRepository.java
@@ -1,8 +1,8 @@
-package com.develhope.spring;
+package com.develhope.spring.User.corrette;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RentRepository extends JpaRepository {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
