+++ src/main/java/com/develhope/spring/User/corrette/UserService.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.User.corrette;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+}
