+++ src/main/java/com/develhope/spring/Vehicles/EntityofVehicles/Car.java
@@ -1,6 +1,7 @@
-package com.develhope.spring;
+package com.develhope.spring.Vehicles.EntityofVehicles;
 
-public class Car extends Vehicles{
+
+public class Car extends Vehicles {
 	public Car(Long id, String type, String brand, String model, Integer displacement, String color, String power, String transmissionType, Integer yearOfRegistration, String fuelType, double price, double discountPrice, String accessories, String condition, Status status) {
 		super(id, type, brand, model, displacement, color, power, transmissionType, yearOfRegistration, fuelType, price, discountPrice, accessories, condition, status);
 	}
