+++ src/main/java/com/develhope/spring/Vehicles/EntityofVehicles/Vans.java
@@ -1,4 +1,4 @@
-package com.develhope.spring;
+package com.develhope.spring.Vehicles.EntityofVehicles;
 
 public class Vans extends Vehicles {
 	public Vans(Long id, String type, String brand, String model, Integer displacement, String color, String power, String transmissionType, Integer yearOfRegistration, String fuelType, double price, double discountPrice, String accessories, String condition, Status status) {
