+++ src/main/java/com/develhope/spring/Vehicles/VehiclesController.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Vehicles;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class VehiclesController {
+}
