+++ src/main/java/com/develhope/spring/Vehicles/VehiclesService.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.Vehicles;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class VehiclesService {
+}
