+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:mysql://127.0.0.1:3306/cars
+    url: jdbc:mysql://127.0.0.1:3306/dealer
     username: root
     password:
   jpa:
